requirements.txt
spotipy
pandas
python-dotenv
pydantic